[BOOKINGSG-9155][GZ] rename styles filename
