+++ src/input-textarea/textarea-counter.styles.ts
@@ -2,9 +2,6 @@ import styled from "styled-components";
 
 import { V3_Colour, V3_Font } from "../v3_theme";
 
-// =============================================================================
-// STYLING
-// =============================================================================
 export const CounterLabel = styled.p`
     text-align: right;
     ${V3_Font["body-sm-semibold"]}
